refatorando codigo e arrumando testes
