+++ lib/menus/books/index.js
@@ -1,16 +1,18 @@
 const actionsModule = require('./library');
 
+const type = 'list';
+const name = 'action';
+const message = 'Menu de livros -- Escolha uma ação';
+
 function run(goBack, dependencies = {}) {
   const { inquirer, library = actionsModule } = dependencies;
 
   return inquirer.prompt({
-    type: 'list',
-    name: 'action',
-    message: 'Menu de livros -- Escolha uma ação',
+    type,
+    name,
+    message,
     choices: [
       { name: 'Pesquisar livros', value: 'list' },
-      { name: 'Pesquisar livro pelo nome', value: 'book' },
-      { name: 'Filtrar livros pelo nome', value: 'listInput'},
       { name: 'Voltar', value: 'back' },
     ],
   })
