+++ lib/menus/books/library/index.js
@@ -1,9 +1,5 @@
 const list = require('./list');
-const book = require('./book');
-const listInput = require('./listInput')
 
 module.exports = {
   list,
-  book,
-  listInput,
 };
