+++ lib/menus/characters/actions/list.js
@@ -6,28 +6,20 @@ async function run(goBack, dependencies, url) {
     const { inquirer, superagent, log = console.log } = dependencies;
 
     const request = superagent.get(url || 'https://www.anapioficeandfire.com/api/characters');
-    // dependendo da string de consulta  o headers.link sofre mudanças;
-    // cria uma string de consulta https://www.anapioficeandfire.com/api/characters?page=2&pageSize=10>
-    // valores depois da ?;
 
     if (!url) request.query({ page: 1, pageSize: 10 });
-    // elementos de consulta vem com esse valor na primeira solicitação;
     request.then(({ body, headers: { link } }) => {
-      // responsável por receber o retorno da solicitação
-      // dependendo da string de consulta  o headers.link sofre mudanças;
-      // console.log(body) // body é tudo o que é retornado, um array de objetos neste caso
-      // console.log(link)
+
       const choices = body.map((character) => {
         const { name, aliases } = character;
 
         return {
           name: name || aliases[0],
-          // se o filme tiver nome, pega o nome, senão pega o apelido que esta em um array;
-          value: character, // recebe todo o objeto;
+          value: character,
         };
       });
 
-      choices.push(new inquirer.Separator()); // serve para fazer uma linha de separação
+      choices.push(new inquirer.Separator());
 
       const links = parseLinks(link);
 
@@ -40,7 +32,6 @@ async function run(goBack, dependencies, url) {
       }
 
       choices.push({ name: 'Voltar para o menu de personagens', value: 'back' });
-      // descobrir a relação do value;
 
       choices.push(new inquirer.Separator());
 
@@ -58,7 +49,7 @@ async function run(goBack, dependencies, url) {
 
           if (!character.url) return resolve();
 
-          log(character); // retorna o objeto de informações;
+          log(character);
 
           return inquirer.prompt({
             type: 'confirm',
