+++ lib/menus/characters/index.js
@@ -13,7 +13,7 @@ function run(goBack, dependencies) {
     ],
   })
     .then(({ action }) => {
-      if (action === 'back') return goBack(); // roda a função que foi passada como parametro do cli, por isso que volta;
+      if (action === 'back') return goBack();
 
       if (actions[action]) {
         return actions[action].run(() => run(goBack, dependencies), dependencies);
