+++ lib/menus/houses/actions/list.js
@@ -1,6 +1,6 @@
 const { parseLinks } = require('../../../utils');
 
-function repeatList(goBack, dependencies, url, resolve) {
+function repeatList(goBack, dependencies, url) {
   const { inquirer } = dependencies;
   return inquirer.prompt({
     type: 'confirm',
@@ -9,8 +9,8 @@ function repeatList(goBack, dependencies, url, resolve) {
     default: true,
   })
     .then(({ repeat }) => {
-      if (!repeat) return resolve();
-      return resolve(run(goBack, dependencies, url));
+      if (!repeat) return;
+      return run(goBack, dependencies, url);
     });
 };
 
@@ -20,7 +20,7 @@ function objectFilter(obj) {
   return objKeysHouses = Object.fromEntries(objKeysHouses);
 };
 
-function listHouses(goBack, dependencies, links, choices, url, resolve) {
+function listHouses(goBack, dependencies, links, choices, url) {
   const { inquirer, log } = dependencies;
   return inquirer.prompt({
     type: 'list',
@@ -31,17 +31,17 @@ function listHouses(goBack, dependencies, links, choices, url, resolve) {
     .then(({ house }) => {
       if (house === 'back') return goBack();
       if (house === 'next' || house === 'prev') {
-        return resolve(run(goBack, dependencies, links[house]));
+        return run(goBack, dependencies, links[house]);
       }
 
-      if (!house.url) return resolve();
+      if (!house.url) return;
 
       log(objectFilter(house));
-      repeatList(goBack, dependencies, url, resolve)
+      repeatList(goBack, dependencies, url)
     });
 };
 
-function chooseOption(choices, link, dependencies, goBack, url, resolve) {
+function chooseOption(choices, link, dependencies, goBack, url) {
   const { inquirer } = dependencies
   choices.push(new inquirer.Separator());
 
@@ -59,10 +59,10 @@ function chooseOption(choices, link, dependencies, goBack, url, resolve) {
 
   choices.push(new inquirer.Separator());
 
-  listHouses(goBack, dependencies, links, choices, url, resolve)
+  listHouses(goBack, dependencies, links, choices, url)
 };
 
-function request(goBack, dependencies, url, resolve) {
+function request(goBack, dependencies, url) {
   const { superagent } = dependencies;
   const request = superagent.get(url || 'https://www.anapioficeandfire.com/api/houses');
 
@@ -78,13 +78,13 @@ function request(goBack, dependencies, url, resolve) {
       };
     });
 
-    chooseOption(choices, link, dependencies, goBack, url, resolve)
+    chooseOption(choices, link, dependencies, goBack, url)
   });
 };
 
 async function run(goBack, dependencies = {}, url) {
   return new Promise((resolve) => {
-    request(goBack, dependencies, url, resolve);
+    request(goBack, dependencies, url);
   });
 };
