+++ lib/menus/houses/actions/list.test.js
@@ -68,14 +68,14 @@ describe('List action', () => {
       await listAction.run(undefined, { inquirer, log, superagent });
     });
 
-    // test('offers to go back to house list after showing details', () => {
-    //   expect(inquirer.prompt).toBeCalledWith({
-    //     type: 'confirm',
-    //     message: 'Deseja consultar outra casa?',
-    //     name: 'repeat',
-    //     default: true,
-    //   });
-    // });
+    test('offers to go back to house list after showing details', () => {
+      expect(inquirer.prompt).toBeCalledWith({
+        type: 'confirm',
+        message: 'Deseja consultar outra casa?',
+        name: 'repeat',
+        default: true,
+      });
+    });
   });
 
   describe('pagination', () => {
